requirements.txt
neo4j==5.14.1
redis==5.0.1
xgboost==2.0.3
fastapi==0.109.0
uvicorn==0.27.0
kafka-python==2.0.2
pandas==2.2.0
pytest==8.0.0
networkx==3.2.1
torch==2.2.0
torch-geometric==2.5.0
